feat: implement pagination for movie list
- Implement pagination using `PagingSource` and `Pager` to load movie data in chunks.
- Create `DiscoverPagingSource` to handle loading data page by page from the API.
- Update `MovieListScreenViewModel` to use `Pager` to fetch and cache movie data.
- Update `MovieListScreen` to use `LazyPagingItems` for displaying paginated data.
- Modify `MoviesRepository` and `MoviesRepositoryImpl` to support fetching movies by page.
- Add loading indicators in `MovieListScreen` while data is being loaded.
- Remove deprecated `MovieListScreenViewModelDELETE`.
- Update `Discover` data class to use `Int` instead of `Long` for page-related properties.

+++ app/src/main/java/com/eaccid/musimpa/entities/Discover.kt
@@ -3,11 +3,11 @@ package com.eaccid.musimpa.entities
 import com.squareup.moshi.Json
 
 data class Discover(
-    val page: Long,
+    val page: Int,
     @Json(name = "results")
     val movies: List<Movie>,
     @Json(name = "total_results")
-    val totalResults: Long,
+    val totalResults: Int,
     @Json(name = "total_pages")
-    val totalPages: Long
+    val totalPages: Int
 )
